# config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Massage Booking API",
    "version": "1.0.0",
    "description": "API documentation for massage booking application"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token"
      }
    },
    "schemas": {
...
      "Massage": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "example": "64a1f2b3c4d5e6f7a8b9c0d2" },
          "name": { "type": "string", "example": "Serene Touch Massage" },
          "address": { "type": "string", "example": "123 Main Street" },
          "district": { "type": "string", "example": "Pathum Wan" },
          "province": { "type": "string", "example": "Bangkok" },
          "postalcode": { "type": "string", "example": "10330" },
          "tel": { "type": "string", "example": "02-123-4567" },
          "picture": { "type": "string", "example": "https://placehold.co/600x400" },
          "price": { "type": "number", "example": 500 },
          "ratingSum": { "type": "number", "example": 45 },
          "userRatingCount": { "type": "number", "example": 10 },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "MassageInput": {
        "type": "object",
        "required": ["name", "address", "district", "province", "postalcode", "tel", "price"],
        "properties": {
          "name": { "type": "string", "example": "Serene Touch Massage" },
          "address": { "type": "string", "example": "123 Main Street" },
          "district": { "type": "string", "example": "Pathum Wan" },
          "province": { "type": "string", "example": "Bangkok" },
          "postalcode": { "type": "string", "example": "10330" },
          "tel": { "type": "string", "example": "02-123-4567" },
          "picture": { "type": "string", "example": "https://placehold.co/600x400" },
          "price": { "type": "number", "example": 500 }
        }
      },
      "Reservation": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "example": "64a1f2b3c4d5e6f7a8b9c0d3" },
          "reserveDate": { "type": "string", "format": "date-time", "example": "2026-03-15T10:00:00.000Z" },
          "user": { "type": "string", "example": "64a1f2b3c4d5e6f7a8b9c0d1" },
          "Massage": { "type": "string", "example": "64a1f2b3c4d5e6f7a8b9c0d2" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "ReservationInput": {
        "type": "object",
        "required": ["reserveDate", "massage"],
        "properties": {
          "reserveDate": { "type": "string", "format": "date-time", "example": "2026-03-15T10:00:00.000Z" },
          "massage": { "type": "string", "description": "ID of the massage shop", "example": "64a1f2b3c4d5e6f7a8b9c0d2" }
        }
      }
    }
  },
  "paths": {
...
    "/api/massages": {
      "get": {
        "summary": "Get all massage shops",
        "tags": ["Massages"],
        "parameters": [
          {
            "in": "query",
            "name": "select",
            "schema": { "type": "string" },
            "description": "Comma-separated fields to include (e.g. name,province)"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": { "type": "string" },
            "description": "Comma-separated fields to sort by (prefix with - for descending)"
          },
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 },
            "description": "Page number for pagination"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 25 },
            "description": "Number of results per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of massage shops",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "count": { "type": "integer" },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "next": {
                          "type": "object",
                          "properties": {
                            "page": { "type": "integer" },
                            "limit": { "type": "integer" }
                          }
                        },
                        "prev": {
                          "type": "object",
                          "properties": {
                            "page": { "type": "integer" },
                            "limit": { "type": "integer" }
                          }
                        }
                      }
                    },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Massage" }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Bad request" }
        }
      },
      "post": {
        "summary": "Create a new massage shop (Admin only)",
        "tags": ["Massages"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/MassageInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Massage shop created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/Massage" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid input" },
          "401": { "description": "Unauthorized - Not logged in" },
          "403": { "description": "Forbidden - Not an admin" }
        }
      }
    },
...
    "/api/reservations": {
      "get": {
        "summary": "Get all reservations (own reservations for users, all for admins)",
        "tags": ["Reservations"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of reservations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "count": { "type": "integer" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Reservation" }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized - Not logged in" },
          "500": { "description": "Cannot find Reservation" }
        }
      },
      "post": {
        "summary": "Create a new reservation (max 3 per user)",
        "tags": ["Reservations"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ReservationInput" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Reservation created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": { "$ref": "#/components/schemas/Reservation" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Massage shop not found, or user already has 3 reservations"
          },
          "401": { "description": "Unauthorized - Not logged in" },
          "500": { "description": "Cannot create Reservations" }
        }
      }
    },
...
  }
}